repo public api refactor
