+++ internal/repodiag/repolog/log_provider.go
@@ -1,0 +1,7 @@
+package repolog
+
+import "github.com/kopia/kopia/internal/repodiag"
+
+type Provider interface {
+	LogManager() *repodiag.LogManager
+}
